Project added
Two TCP connections established upon playing game, these connect to the two servers created in the python main script. One handles position of the drone, the other allows for the transfer of the drone camera feed for use with the vision algorithm.
Currently a follow script on the drone cam is being tested with, it may not be permanent.